FIX review lawyer

+++ src/app/(dashboard)/lead-management/LeadManagement.tsx
@@ -140,7 +140,7 @@ const LeadManagement = () => {
       //     status: 'IN PROGRESS',
       //   };
       //   await database.updateData(
-      //     `${process.env.NEXT_PUBLIC_URL_LEADS}/${filterData[index]['lead id']}`,
+      //     `${process.env.NEXT_PUBLIC_URL}/leads/${filterData[index]['lead id']}`,
       //     dataUpdate
       //   );
       //   fetchLeads();
@@ -159,7 +159,7 @@ const LeadManagement = () => {
     }
     if (selectedChange === 'LOST') {
       const deleteAssined = await database.deleteData(
-        `${process.env.NEXT_PUBLIC_URL_LEADS_ASSIGNED}/lead/${selectedLead['lead id']}`
+        `${process.env.NEXT_PUBLIC_URL}/leads-assigned/lead/${selectedLead['lead id']}`
       );
       if (!deleteAssined.success) {
         return toast.error('Error to delete lawyer');
@@ -168,14 +168,15 @@ const LeadManagement = () => {
     setloading(true);
 
     const dataUpdate = {
-      status:
-        e.target.checkbox.checked === true ? 'DISABLED' : e.target.status.value,
-      comments: e.target.comments.value,
+      // status:
+      //   e.target.checkbox.checked === true ? 'DISABLED' : e.target.status.value,
+      status: e.target.status.value,
+      comments: e.target.status.value === 'NEW' ? '' : e.target.comments.value,
       description: e.target.description.value,
     };
 
     const responseUpdate = await database.updateData(
-      `${process.env.NEXT_PUBLIC_URL_LEADS}/${selectedLead['lead id']}`,
+      `${process.env.NEXT_PUBLIC_URL}/leads/${selectedLead['lead id']}`,
       dataUpdate
     );
     if (!responseUpdate.success) {
@@ -188,11 +189,23 @@ const LeadManagement = () => {
     setloading(false);
   };
   const DeleteLead = async () => {
+    console.log(selectedLead.status);
+
+    if (
+      selectedLead.status !== 'NEW' &&
+      selectedLead.status !== 'DISABLED' &&
+      selectedLead.status !== 'LOST'
+    ) {
+      return toast.error(
+        'You cannot delete the lead because it is assigned to a lawyer.'
+      );
+    }
+
     const dataDelete = await database.deleteData(
-      `${process.env.NEXT_PUBLIC_URL_LEADS}/${selectedLead['lead id']}`
+      `${process.env.NEXT_PUBLIC_URL}/leads/${selectedLead['lead id']}`
     );
     const deleteAssined = await database.deleteData(
-      `${process.env.NEXT_PUBLIC_URL_LEADS_ASSIGNED}/lead/${selectedLead['lead id']}`
+      `${process.env.NEXT_PUBLIC_URL}/leads-assigned/lead/${selectedLead['lead id']}`
     );
     if (!dataDelete.success || !deleteAssined.success) {
       return toast.error('Error to delete lawyer');
@@ -222,9 +235,9 @@ const LeadManagement = () => {
     return <ReLoading />;
   }
   return (
-    <div className="flex flex-col gap-5">
-      <Modal title="Lead info" isOpen={isOpenLead} setIsOpen={setIsOpenLead}>
-        <div className="px-8">
+    <div className='flex flex-col gap-5'>
+      <Modal title='Lead info' isOpen={isOpenLead} setIsOpen={setIsOpenLead}>
+        <div className='px-8'>
           <p>
             Selection date :{' '}
             {Object.keys(selectedLead).length > 0 &&
@@ -237,16 +250,16 @@ const LeadManagement = () => {
             <CountdownTimer targetDate={selectedLead['date_updated']} />
           )}
 
-          <p className="text-red-500">
+          <p className='text-red-500'>
             This lead will be marked as lost and will not be reinstated.
           </p>
-          <p className="text-4xl py-6 ">{selectedLead?.['lead name']}</p>
-          <form onSubmit={saveLeadContact} className="grid grid-cols-3 gap-2 ">
-            <p className="">Status:</p>
+          <p className='text-4xl py-6 '>{selectedLead?.['lead name']}</p>
+          <form onSubmit={saveLeadContact} className='grid grid-cols-3 gap-2 '>
+            <p className=''>Status:</p>
 
             <Input
-              type="select"
-              name="status"
+              type='select'
+              name='status'
               values={
                 selectedLead.status === 'DISABLED' ||
                 selectedLead.status === 'LOST'
@@ -267,50 +280,50 @@ const LeadManagement = () => {
             />
             <p></p>
             <p>Email:</p>
-            <p className="col-span-2 text-gray-500 ">{selectedLead?.email}</p>
+            <p className='col-span-2 text-gray-500 '>{selectedLead?.email}</p>
             <p>Phone number:</p>
-            <p className="col-span-2 text-gray-500">
+            <p className='col-span-2 text-gray-500'>
               {selectedLead?.['phone number']}
             </p>
             <p>Service Type:</p>
-            <p className="col-span-2 text-gray-500">{selectedLead?.service}</p>
+            <p className='col-span-2 text-gray-500'>{selectedLead?.service}</p>
             <p>Description:</p>
             <textarea
-              name="description"
-              className="col-span-2 text-gray-500 border-2 bg-gray-100 rounded-md"
-              placeholder=" Leave your comment....."
+              name='description'
+              className='col-span-2 text-gray-500 border-2 bg-gray-100 rounded-md'
+              placeholder=' Leave your comment.....'
             >
               {selectedLead?.['description lead']}
             </textarea>
 
             <p>Comment:</p>
             <textarea
-              name="comments"
-              className="col-span-2 text-gray-500 border-2 bg-gray-100 rounded-md"
-              placeholder=" Leave your comment....."
+              name='comments'
+              className='col-span-2 text-gray-500 border-2 bg-gray-100 rounded-md'
+              placeholder=' Leave your comment.....'
             >
               {selectedLead?.comments}
             </textarea>
             <p></p>
-            <p className="flex gap-1 col-span-2 text-gray-500 ">
+            {/* <p className='flex gap-1 col-span-2 text-gray-500 '>
               <input
-                name="checkbox"
+                name='checkbox'
                 id={`checkbox-lead`}
-                className="peer hidden"
-                type="checkbox"
+                className='peer hidden'
+                type='checkbox'
               />
               <label
                 htmlFor={`checkbox-lead`}
-                className="flex items-center justify-center w-5 h-5 border border-green-500 rounded bg-white cursor-pointer relative  text-white peer-checked:text-green-500"
+                className='flex items-center justify-center w-5 h-5 border border-green-500 rounded bg-white cursor-pointer relative  text-white peer-checked:text-green-500'
               >
-                <i className="fi fi-rr-check absolute  text-sm  peer-checked:block "></i>
+                <i className='fi fi-rr-check absolute  text-sm  peer-checked:block '></i>
               </label>{' '}
               Not contact this lead again
-            </p>
-            <div className="col-end-4 text-right">
+            </p> */}
+            <div className='col-end-4 text-right'>
               <Button
-                name="Save"
-                type="submit"
+                name='Save'
+                type='submit'
                 disabled={loading}
                 color={`${
                   loading ? 'bg-gray-500 animate-pulse' : 'bg-primary'
@@ -318,42 +331,42 @@ const LeadManagement = () => {
               />
             </div>
           </form>
-          <p className="text-red-500 text-sm py-4">
+          <p className='text-red-500 text-sm py-4'>
             The super admin will review this case, leave us a clear comment.
           </p>
         </div>
       </Modal>
       <Modal
-        title="Delete"
+        title='Delete'
         isOpen={isOpenDelete}
         setIsOpen={setIsOpenDelete}
-        className="max-w-sm"
+        className='max-w-sm'
       >
-        <div className="flex flex-col gap-4">
-          <div className="flex justify-center text-center">
+        <div className='flex flex-col gap-4'>
+          <div className='flex justify-center text-center'>
             <p>
               Are you sure you want to delete the lead{' '}
-              <span className="font-medium">{selectedLead?.email}?</span>
+              <span className='font-medium'>{selectedLead?.email}?</span>
             </p>
           </div>
 
-          <div className="flex justify-around">
+          <div className='flex justify-around'>
             <Button
-              name="Cancel"
-              type="button"
+              name='Cancel'
+              type='button'
               onClick={() => setIsOpenDelete(false)}
             />
             <Button
-              name="Delete"
-              type="button"
-              color="bg-red-500"
+              name='Delete'
+              type='button'
+              color='bg-red-500'
               onClick={DeleteLead}
             />
           </div>
         </div>
       </Modal>
-      <Tilte name="Lead Management" search={true} filterSearch={filterSearch} />
-      <div className="flex gap-2 flex-wrap">
+      <Tilte name='Lead Management' search={true} filterSearch={filterSearch} />
+      <div className='flex gap-2 flex-wrap'>
         <button
           onClick={() => handleStatusClick(null)}
           className={`px-4 p-1 rounded text-sm ${
